Waste-Classification-using-YOLOv8/requirements.txt
# Core ML and Computer Vision
ultralytics>=8.0.0
torch>=1.7.0
torchvision>=0.8.1
opencv-python>=4.6.0
Pillow>=7.1.2

# Streamlit and Web Interface
streamlit>=1.28.0

# Data Processing
numpy>=1.18.5
pandas>=1.1.4
matplotlib>=3.2.2
seaborn>=0.11.0

# Utilities
tqdm>=4.64.0
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
psutil

# Optional for video processing
pafy
youtube-dl

# For model serialization
pickle-mixin
